# helpan-ai-openapi-v1.yaml
openapi: 3.1.0
info:
  title: Helpan AI Rail
  version: '1.0.0'
  summary: Agent rail for the Kirimon Market Ventures platform — orchestration runtime, delegated authority, consent surface, and audit for AI agents acting on behalf of users across Kipkiren Pay, Identiti, and Todoku.
  description: |
    Helpan AI is the **fourth platform rail** in the Kirimon Market Ventures (KMV) stack, alongside
    Identiti (identity), Kipkiren Pay (payment), and Todoku (comms). It provides the agent orchestration
    runtime, briefing storage, matching engine, delegated authority issuance/validation/revocation, the
    Helpan Console consent surface, audit log, and OAuth scope catalogue for third-party agents.

    Per design law (Helpan AI Instruction Pack §3) and the Agentic AI Signal Scan integration (7 May 2026):
    - Helpan AI does NOT hold funds, send messages, or store identity material. Every primitive that
      belongs to KP, Identiti, or Todoku is invoked via the relevant rail's API.
    - Every agent action carries a scoped, time-bounded, revocable **delegated authority token**.
    - Every issued delegated authority is visible in the Helpan Console with one-tap revocation.
    - Per-call validation of delegated authorities against this rail's `POST /authorities/{id}/validate`
      endpoint is mandatory for relying parties (KP, Todoku, consuming-app servers).
    - `initiated_by` and `actor` claims propagate cross-rail per Schema Appendix amendments §A on
...

...
  contact:
...
servers:
...
security:
  - HmacSha256: []
  - BearerCustomer: []
  - BearerService: []
paths:
...
  /operator/agents:
    post:
      tags: [Operator]
      summary: Register an agent (consuming-app branded agent or third-party)
      operationId: registerAgent
      security:
        - HmacSha256: [helpan:admin]
      parameters:
        - $ref: '#/components/parameters/XIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterAgentRequest'
      responses:
        '201':
          description: Agent registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
  /operator/agents/{agent_id}:
    parameters:
      - $ref: '#/components/parameters/AgentId'
    get:
      tags: [Operator]
      summary: Read agent detail (operator)
      operationId: getAgent
      security:
        - HmacSha256: [helpan:admin]
      responses:
        '200':
          description: Agent detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'

    patch:
      tags: [Operator]
      summary: Update agent status (suspend/reactivate)
      operationId: updateAgentStatus
      security:
        - HmacSha256: [helpan:admin]
      parameters:
        - $ref: '#/components/parameters/XIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAgentStatusRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
  /operator/safety-policies:
    get:
      tags: [Operator]
      summary: List per-app safety policies
      operationId: listSafetyPolicies
      security:
        - HmacSha256: [helpan:admin]
      responses:
        '200':
          description: Safety policies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SafetyPolicyListResponse'
  /operator/safety-policies/{policy_id}:
    parameters:
      - $ref: '#/components/parameters/PolicyId'
    put:
      tags: [Operator]
      summary: Update per-app safety policy (category whitelist, content moderation rules)
      operationId: updateSafetyPolicy
      security:
        - HmacSha256: [helpan:admin]
      parameters:
        - $ref: '#/components/parameters/XIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SafetyPolicy'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SafetyPolicyResponse'
...
components:
...
  securitySchemes:
    HmacSha256:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Format: `Helpan-HMAC-SHA256 app_id={app_id}, signature={base64_sig}`.
        Canonical string (identical across all four rails per App Integration Guide §5.7):
        `{METHOD}\n{PATH_AND_QUERY}\n{CONTENT_TYPE}\n{X-Helpan-Timestamp}\n{SHA256_HEX(body)}`.
        mTLS additionally required at the transport layer.

    BearerCustomer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Identiti-issued customer JWT. Audience must include this rail's audience
        (`https://api.helpan.co.ke`). Carries `sub` (Account UUID), `tier`, `actor`,
        `initiated_by`, `auth_factors`, `exp`. Per Identiti Rail Contract Amendment §A.4:
        30-min standard / 5-min elevated TTL.

    BearerService:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Identiti-issued service JWT for service-to-service authentication
        (consuming-app servers calling Helpan AI on behalf of the user). 15-min standard /
        5-min elevated TTL per Identiti Rail Contract Amendment §A.4.

  # --------------------------------------------------------------------------
  # PARAMETERS
  # --------------------------------------------------------------------------
  parameters:
...
    XIdempotencyKey:
      name: X-Idempotency-Key
      in: header
      required: true
      schema: { type: string, format: uuid }
      description: |
        UUIDv4 per logical operation. Apps SHOULD use the UUIDv5-from-business-op-id pattern
        per App Integration Guide §7.6. 24-hour idempotency window.
...
    AgentId:
      name: agent_id
      in: path
      required: true
      schema: { type: string, pattern: '^agt_[0-9A-HJKMNP-TV-Z]{26}$' }
    PolicyId:
      name: policy_id
      in: path
      required: true
      schema: { type: string, pattern: '^sfp_[0-9A-HJKMNP-TV-Z]{26}$' }

  # --------------------------------------------------------------------------
  # SCHEMAS
  # --------------------------------------------------------------------------
  schemas:
...
    Meta:
      type: object
      required: [request_id, timestamp, schema_version]
      properties:
        request_id:
          type: string
          description: ULID
        timestamp:
          type: string
          format: date-time
        schema_version:
          type: string
          enum: ['1.0']
    SuccessEnvelope:
      type: object
      required: [ok, data, meta]
      properties:
        ok:
          type: boolean
          enum: [true]
        data:
          description: Operation-specific payload
        meta:
          $ref: '#/components/schemas/Meta'
...
    AgentClass:
      type: string
      enum: [portfolio_app, third_party_oauth, internal_system]
    AgentStatus:
      type: string
      enum: [active, suspended, retired]
    Agent:
      type: object
      required: [id, name, agent_class, status, created_at]
      properties:
        id:
          type: string
          pattern: '^agt_[0-9A-HJKMNP-TV-Z]{26}$'
        name:
          type: string
        agent_class:
          $ref: '#/components/schemas/AgentClass'
        owner_app_id:
          type: string
          description: For portfolio_app agents — the consuming app that ships this agent
        third_party_oauth_client_id:
          type: string
          description: For third_party_oauth agents — the OAuth client_id registered with Identiti
        status:
          $ref: '#/components/schemas/AgentStatus'
        created_at:
          type: string
          format: date-time
        suspended_at:
          type: string
          format: date-time
          nullable: true
    RegisterAgentRequest:
      type: object
      required: [name, agent_class]
      properties:
        name: { type: string }
        agent_class: { $ref: '#/components/schemas/AgentClass' }
        owner_app_id: { type: string }
        third_party_oauth_client_id: { type: string }
    UpdateAgentStatusRequest:
      type: object
      required: [status]
      properties:
        status: { $ref: '#/components/schemas/AgentStatus' }
        reason: { type: string }
    AgentResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/Agent' }

    # ---- safety policies --------------------------------------------------
    SafetyPolicy:
      type: object
      required: [id, app_id]
      properties:
        id:
          type: string
          pattern: '^sfp_[0-9A-HJKMNP-TV-Z]{26}$'
        app_id:
          type: string
        category_whitelist:
          type: array
          items: { type: string }
        category_blacklist:
          type: array
          items: { type: string }
        content_moderation_rules:
          type: array
          items:
            type: object
            properties:
              rule_id: { type: string }
              kind:
                type: string
                enum: [text_filter, image_filter, link_filter, recipient_filter]
              pattern: { type: string }
              action:
                type: string
                enum: [reject, flag, redact]
        audience_posture:
          type: string
          enum: [family_friendly, general, adult_confirmed]
        location_precision_floor:
          type: string
          enum: [merchant_level, neighbourhood_level, city_level, none]
        updated_at:
          type: string
          format: date-time
    SafetyPolicyListResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                items:
                  type: array
                  items: { $ref: '#/components/schemas/SafetyPolicy' }
    SafetyPolicyResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/SafetyPolicy' }

    # ---- audit log --------------------------------------------------------
...